Fix packaging files: RPM spec, appdata, metainfo
Update the image element in the metadata to an existing, freedesktop-compliant
screenshot for Appstream-compliant software stores (GNOME Software, KDE
Software, etc). This screenshot should be displayed by these software store
applications when the Flatpak is being installed by end-users, along with
developer, license, source-code availability, and other metadata that may
interest users.

Include the missing changelog entry in the RPM spec for the release of v4.0.
Include another entry in the changelog to signify that fix, and increment the
release of the RPM package version. TODO: trigger a rebuild of the RPMs on COPR,
and rebuild the openSUSE Leap 15.4 RPM locally and upload the new RPM to
Packagecloud.io.