Insert RPM spec changelog entry, increment verison in metainfo
Ben incremented the version tag in the spec file, and indicated this is the first release of the RPM in the relese tag in the spec file, but forgot to add a changelog entry. I added an appropriate changelog entry for this version of SLiM (and its RPM) on behalf of Ben.

I incremented the version attribute of the release tag in the Appstream/Flatpak Metainfo XML.