Specify source and build directory manually on RHEL 8
RHEL 8 does not, apparently, specify source and build directories when it calls `cmake` (not `cmake --build`, that's a different step). This causes CMake to set the source and build directories to both be the current directory; apparently this causes the macros `%_vpath_srcdir` and `%_vpath_builddir` to be ignored during `cmake --build`, which may be the ultimate cause of the issues we've encountered recently with the CMake build system on RHEL 8 builders.

We'll see how this commit affects the RHEL 8 builders. It shouldn't affect the others, because when the builder is not RHEL 8 the macro `%sourceAndBuilddir` should expand to an empty string.